backend_django/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-countries==7.6.1
django-filter==24.3
django-js-asset==3.1.2
django-redis==5.4.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mercadopago==2.3.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
psutil==7.0.0
psycopg==3.2.4
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.1
